changed bignum.c to match the struct in bignum.h

+++ src/bignum.c
@@ -1,5 +1,5 @@
 #include "bignum.h"
-#include <cstdint>
+#include <stdio.h>
 #include <stddef.h>
 #include <stdint.h>
 #include <stdlib.h>
@@ -21,26 +21,23 @@ void bn_free(bn_t *n){
 }
 
 int bn_add(bn_t *r, const bn_t *a, const bn_t *b){
-    uint32_t len = a -> len ;
+    size_t len = a -> len ; // for now assume that a and b have the same length
     uint32_t carry = 0;
-
 
     if (r-> capacity < len +1)
         return -1;
 
-    for (int i = 0; i < len; i++){
+    for (size_t i = 0; i < len; i++){
 
         uint64_t temp_sum = (uint64_t) a -> digits[i] + b -> digits[i] + carry;
         uint32_t result_word = (uint32_t) temp_sum;
         carry = temp_sum >> 32;
         r -> digits[i] = result_word;
-
     }
     r-> len = len;
     if (carry != 0){
         r-> digits[len] = carry;
         r-> len++;
     }
-
     return 0;
 }
